Task 2: Debug StaticResource key
I got an AnonymusMethod ( UnhandledException )  error at runtime. The reason for this that the static resources are mostly resolved in parse time, but it does not check the corresponding keys in build time. The parsing of XAML is made in two steps. The first step in build time converts the XAML to a BAML file, a binary optimized version of the layout. The final parsing of the XAML is done in the runtime, where all component linked together. Then the application finds out that the resource with this key is not defined, so it throws an exception.