Task 3: Debug GameViewModel SelectedField property
The "vm.SelectedField=current;" line do the same thing as the "current.IsSelected=true;", but the first one is a bit more. Not only selects the current field, but makes sure, that the previously selected field gets deselected, which means, the player can change his mind. Furthermore, the Execute method later checks whether the SelectedField property is null, which will be always null if we only use the IsSelected property.